fix(vimwiki): `VimwikiHeader` highlights and URL metadata (#575)
## What is this PR for?
Vimwiki support was added in
[8850021bed1d8595581a5903fd03acacf0d12af5](https://github.com/folke/tokyonight.nvim/commit/8850021bed1d8595581a5903fd03acacf0d12af5)
but there was a typo in the generation of the `VimwikiHeader` highlights
and instead they were generated under `VimwikiHeaer`. It also seems the
the URL was set to headlines.nvim by mistake, presumably because this
file was based off of the headlines.nvim group file.

Disclaimer: I don't actually use vimwiki but I was looking through the
commits and saw this easy to fix bug so I did it myself.