feat(extra): add Tailwind CSS v4 (#701)
## Description

<!-- Describe the big picture of your changes to communicate to the
maintainers
  why we should accept this pull request. -->

Adds all theme variations as CSS variables, under the `@theme`
directive. Allows for use in a project using Tailwind CSS v4. Uses
`oklch` for the color definitions, using the `from` feature to use
existing hex colors.

## Related Issue(s)

<!--
  If this PR fixes any issues, please link to the issue here.
  - Fixes #<issue_number>
-->

N/A. I was going to open an issue for this, but I had already done this
for my personal use and figured I would chuck it upstream in case you
are interested in bringing it in.

## Screenshots

<!-- Add screenshots of the changes if applicable. -->

Demos of the themes on Tailwind's Play web playground:

- [tokyonight_day](https://play.tailwindcss.com/kHihIJmxtf)
- [tokyonight_moon](https://play.tailwindcss.com/8yI7iYjuFj)
- [tokyonight_night](https://play.tailwindcss.com/L91QMbz7XU)
- [tokyonight_storm](https://play.tailwindcss.com/9mZNmRRIVQ)


![Screenshot_20250324_145218](https://github.com/user-attachments/assets/dfa53ad3-2312-4523-8f66-a95430644cad)