fix(extra): cleanup alacritty theme (#704)
This PR cleans up a couple of issues with the Alacritty color scheme:

1. Removes commented-out cursor color. It's not used (obviously :grin:
). But also the commented out values duplicate the default Alacritty
behavior of using the fg and bg reversed.

2. Fixes indexed colors 16 and 17. I show that xterm 256‑color standards
([source](https://www.ditig.com/256-colors-cheat-sheet)) indicate 16 and
17 should be Grey0 and NavyBlue, not orange and red. I'd be curious to
know if there was some reason for mapping these two specific colors to
orange and red. But it seems like that information may be lost to time.
I went back in git history and I couldn't find any notes. I noticed this
issue because this was causing the ngrok tool to render illegibly, since
it apparently uses 16 as it's background color.